Automatically run testcrate in root tests
