+++ Cargo.toml
@@ -17,3 +17,6 @@ members = ["testcrate"]
 [dependencies]
 cc = { version = "1", features = ["parallel"] }
 dircpy = "0.3.8"
+
+[dev-dependencies]
+testcrate = { path = "./testcrate", features = ["libsodium"] }
